added api for addressdetails and getall , get by id , delete api for hospital service.

+++ SIFO/SIFO.Model/Response/CityResponse.cs
@@ -8,7 +8,7 @@
         public string? StateCode { get; set; }
         public string? StateName { get; set; }
         public string? CountryName { get; set; }
-        public int? CountryId { get; set; }
+        public long? CountryId { get; set; }
         public string? CountryCode { get; set; }
         public decimal? Latitude { get; set; }
         public decimal? Longitude { get; set; }
