+++ SIFO/SIFO.Model/Response/StateResponse.cs
@@ -4,7 +4,7 @@
     {
         public long Id { get; set; }
         public string Name { get; set; }
-        public int? CountryId { get; set; }
+        public long? CountryId { get; set; }
         public string CountryCode { get; set; }
         public string CountryName { get; set; }
         public string FipsCode { get; set; }
